refactor(dashboard): split inbox-sections into focused modules
inbox-sections.tsx had grown to ~1.7k lines bundling the data layer,
drawer presentation, message rendering, and the activity rail. Extract:

- inbox-types.ts          — shared DTO/controller types
- inbox-helpers.ts        — pure data helpers (merge, truncate, etc.)
- inbox-data.ts           — useInboxData hook + queue builder
- inbox-message-bubble.tsx — MessageBubble + markdown rendering
- inbox-activity-rail.tsx  — ActivityRail
- inbox-conv-drawers.tsx   — Simplified/Full conv drawers + load-failed

inbox-sections.tsx keeps the section/list components (LiveNowSection,
QueueSection, InboxDrawers, LiveCard, QueueRow) and re-exports
useInboxData / InboxController / ConvActionError / QueueItem so existing
importers are unchanged. Pure code movement, no behavior change.
